Fix some quality flaws
